LayoutTests/media/media-can-play-h265-video.html

<!DOCTYPE html>
<html>
    <head>
        <script src="video-test.js"></script>
        <script>
            function start()
            {
                video = document.getElementsByTagName('video')[0];

                testExpected("video.canPlayType('video/mp4')", "maybe");
                testExpected("video.canPlayType('video/mp4; Codecs=\"hvc1.2.20000000.l123.b0\"')", "probably");
                testExpected("video.canPlayType('        Video/MP4 ; CODECS=\"hev1.1.6.L150.B0\"')", "probably");

                endTest();
            }
        </script>
    </head>
    <body onload="start()">
        <video controls></video>
        <p>Test HTMLMediaElement <em>canPlayType()</em> method with multiple video mpeg4 MIME types with H265 codecs.</p>
        <p>These tests may be expected to fail if the WebKit port does not support the format.</p>
    </body>
</html>
